requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
imageio==2.34.0
imutils==0.5.4
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
threadpoolctl==3.3.0
tifffile==2024.2.12
